+++ backend/src/docpool.base/docpool/base/marker.py
@@ -5,3 +5,9 @@ class IImportingMarker(Interface):
     """
     A marker on the request used during importing with exportimport.
     """
+
+
+class IAppActiveMarker(Interface):
+    """
+    Allow querying for app-active marker interfaces by app name.
+    """